started work on ic2 classic compat
