+++ src/main/java/com/rumaruka/emt/block/solar/compressed/BlockDoubleCompressedSolar.java
@@ -1,6 +1,5 @@
 package com.rumaruka.emt.block.solar.compressed;
 
-import com.rumaruka.emt.tile.solar.compressed.TileEntityCompressSolar;
 import com.rumaruka.emt.tile.solar.compressed.TileEntityDoubleCompressedSolar;
 import net.minecraft.block.BlockContainer;
 import net.minecraft.block.BlockHorizontal;
